requirements.txt
numpy == 1.19.2
cupy-cuda101==8.0.0